LayoutTests/platform/chromium-win-xp/inspector/extensions/extensions-resources-expected.txt

CONSOLE MESSAGE: line 12: don't panic!
Tests resource-related methods of WebInspector extension API


Started extension.
Running tests...
RUNNING TEST: extension_testGetAllResources
page resources:
{
    0 : {
        0 : {
            getContent : <function>
            setContent : <function>
            type : "image"
            url : .../abe.png
        }
        1 : {
            getContent : <function>
            setContent : <function>
            type : "stylesheet"
            url : .../audits-style1.css
        }
        2 : {
            getContent : <function>
            setContent : <function>
            type : "document"
            url : .../extensions-resources.html
        }
        3 : {
            getContent : <function>
            setContent : <function>
            type : "script"
            url : .../extensions-test.js
        }
        4 : {
            getContent : <function>
            setContent : <function>
            type : "script"
            url : .../inspector-test.js
        }
        5 : {
            getContent : <function>
            setContent : <function>
            type : "script"
            url : .../inspector-test.js
        }
        6 : {
            getContent : <function>
            setContent : <function>
            type : "script"
            url : .../test-script.js
        }
    }
}
RUNNING TEST: extension_testGetResourceContent
{
    0 : "function __test_function()
{
}

"
    1 : ""
}
RUNNING TEST: extension_testOnContentCommitted
log: Extension server error: Object does not support requested operation: Resource is not editable
content committed for resource .../audits-style1.css (type: stylesheet), new content: div.test { width: 220px; height: 42px; }
Revision content: div.test { width: 220px; height: 42px; }
RUNNING TEST: extension_testOnResourceAdded
Debugger was enabled.
resource added:
{
    0 : {
        getContent : <function>
        setContent : <function>
        type : "script"
        url : data:application/javascript...
    }
}
Debugger was disabled.
RUNNING TEST: extension_testOpenResourceHandler
handleOpenResource() invoked [this should only appear once!]: 
{
    0 : {
        getContent : <function>
        setContent : <function>
        type : "document"
        url : .../extensions-resources.html
    }
    1 : 12
}
RUNNING TEST: extension_testSetResourceContent
div.test width after stylesheet edited (should be 126): 126
All tests done.
